Update 3
Fixed some bugs that were happening with the ScoreKeeper. On the very first goal of the instance, the game would not end if scoreToWin was set to 1. There was also some singleton logic issues that were fixed as well.

Added:
Main Menu
Game Over Menu

Still Needed:
Basic AI logic for player 2