Final Update
Added:

Basic AI for player 2.

Collision/Bounds for player and AI so that they can't leave the "court."

Fixed:

On very rare occasions if you hit a ball in between a paddle and wall you could cause it to fly out of bounds, effectively breaking the game. Changed the logic to only allow a paddle to hit the ball once, after the other player paddle has hit it.

This is the final update, the only things that would ever need to be changed/added would be a higher level/smarter AI, or to adjust the amount of points needed to win the game. Currently set default is one point to win.

Otherwise this is complete, and hopefully bug free.